requirements.txt
asgiref==3.4.1
class-registry==2.1.2
dj-database-url==0.5.0
Django==3.2.8
django-ckeditor==6.1.0
django-ckeditor-5==0.0.14
django-ckeditor5==0.0.4
django-crispy-forms==1.13.0
django-etc==1.3.0
django-filter==21.1
django-heroku==0.3.1
django-hitcount==1.3.3
django-js-asset==1.2.2
filters==1.3.2
filters-django==1.0.5
Pillow==8.4.0
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.23
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
